Revert "clean up code, remove unnecessary async"
This reverts commit 0e039c41fe01996160bee75aa9de775dc59066ce.
